updated storage to 1d array

+++ Python/wrap.py
@@ -4,7 +4,7 @@ import ctypes
 import numpy as np
 from past.types import long
 
-from Ising import DoubleVector
+from Ising import IntVector
 from Python.Ising import vector_data
 
 
@@ -14,7 +14,7 @@ def PtrToArray(ptr, size):
     return np.copy(np.ctypeslib.as_array(
         (ctypes.c_int32 * size).from_address(addr)))
 
-def VecToMat(vec : DoubleVector):
+def VecToMat(vec : IntVector):
     data = vector_data(vec)
     array = PtrToArray(data.ptr, data.size)
